Moved logging-related functions to Logging and functions for generating specific modules to Modules
