+++ code/drasil-code/drasil-code.cabal
@@ -31,6 +31,8 @@ library
     , Language.Drasil.Code.Imperative.LanguageRenderer.PythonRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.CSharpRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.CppRenderer
+    , Language.Drasil.Code.Imperative.Logging
+    , Language.Drasil.Code.Imperative.Modules
     , Language.Drasil.Code.Imperative.Parameters
     , Language.Drasil.Code.CodeGeneration
     , Language.Drasil.CodeSpec
@@ -87,6 +89,8 @@ executable codegenTest
     , Language.Drasil.Code.Imperative.LanguageRenderer.PythonRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.CSharpRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.CppRenderer
+    , Language.Drasil.Code.Imperative.Logging
+    , Language.Drasil.Code.Imperative.Modules
     , Language.Drasil.Code.Imperative.Parameters
     , Language.Drasil.Code.CodeGeneration
     , Language.Drasil.CodeSpec
